Lease test re-enabled and passing
